refactor(combo): extract auto strategy to combo/autoStrategy.ts
